+++ llvm/lib/ObjectYAML/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_llvm_component_library(LLVMObjectYAML
   ArchiveEmitter.cpp
   ArchiveYAML.cpp
+  BBAddrMapYAML.cpp
   CodeViewYAMLDebugSections.cpp
   CodeViewYAMLSymbols.cpp
   CodeViewYAMLTypeHashing.cpp
