+++ llvm/tools/obj2yaml/elf2yaml.cpp
@@ -892,9 +892,9 @@ ELFDumper<ELFT>::dumpBBAddrMapSection(const Elf_Shdr *Shdr) {
   unsigned AddressSize = ELFT::Is64Bits ? 8 : 4;
   DataExtractor Data(Content, Obj.isLE());
 
-  std::vector<ELFYAML::BBAddrMapEntry> Entries;
+  std::vector<BBAddrMapYAML::BBAddrMapEntry> Entries;
   bool HasAnyPGOAnalysisMapEntry = false;
-  std::vector<ELFYAML::PGOAnalysisMapEntry> PGOAnalyses;
+  std::vector<BBAddrMapYAML::PGOAnalysisMapEntry> PGOAnalyses;
   DataExtractor::Cursor Cur(0);
   uint8_t Version = 0;
   uint16_t Feature = 0;
@@ -921,7 +921,7 @@ ELFDumper<ELFT>::dumpBBAddrMapSection(const Elf_Shdr *Shdr) {
       Address = Data.getUnsigned(Cur, AddressSize);
       NumBlocks = Data.getULEB128(Cur);
     }
-    std::vector<ELFYAML::BBAddrMapEntry::BBRangeEntry> BBRanges;
+    std::vector<BBAddrMapYAML::BBAddrMapEntry::BBRangeEntry> BBRanges;
     uint64_t BaseAddress = 0;
     for (uint64_t BBRangeN = 0; Cur && BBRangeN != NumBBRanges; ++BBRangeN) {
       if (FeatureOrErr->MultiBBRange) {
@@ -931,7 +931,7 @@ ELFDumper<ELFT>::dumpBBAddrMapSection(const Elf_Shdr *Shdr) {
         BaseAddress = Address;
       }
 
-      std::vector<ELFYAML::BBAddrMapEntry::BBEntry> BBEntries;
+      std::vector<BBAddrMapYAML::BBAddrMapEntry::BBEntry> BBEntries;
       // Read the specified number of BB entries, or until decoding fails.
       for (uint64_t BlockIndex = 0; Cur && BlockIndex < NumBlocks;
            ++BlockIndex) {
@@ -960,7 +960,7 @@ ELFDumper<ELFT>::dumpBBAddrMapSection(const Elf_Shdr *Shdr) {
     Entries.push_back(
         {Version, Feature, /*NumBBRanges=*/{}, std::move(BBRanges)});
 
-    ELFYAML::PGOAnalysisMapEntry &PGOAnalysis = PGOAnalyses.emplace_back();
+    BBAddrMapYAML::PGOAnalysisMapEntry &PGOAnalysis = PGOAnalyses.emplace_back();
     if (FeatureOrErr->hasPGOAnalysis()) {
       HasAnyPGOAnalysisMapEntry = true;
